+++ include/slug_ffi_prototype.h
@@ -6,7 +6,7 @@
 #include <stdint.h>
 
 #define SLUG_FFI_PROTOTYPE_ABI_MAJOR 0u
-#define SLUG_FFI_PROTOTYPE_ABI_MINOR 1u
+#define SLUG_FFI_PROTOTYPE_ABI_MINOR 2u
 
 typedef enum {
   SLUG_FFI_OK = 0,
@@ -23,9 +23,12 @@ typedef struct {
 
 typedef bool (*slug_ffi_argument_i64_fn)(slug_ffi_call *, size_t, int64_t *);
 typedef bool (*slug_ffi_argument_f64_fn)(slug_ffi_call *, size_t, double *);
+typedef bool (*slug_ffi_argument_resource_fn)(slug_ffi_call *, size_t, slug_ffi_text, void **);
 typedef void (*slug_ffi_set_i64_fn)(slug_ffi_call *, int64_t);
 typedef void (*slug_ffi_set_f64_fn)(slug_ffi_call *, double);
 typedef void (*slug_ffi_set_error_fn)(slug_ffi_call *, slug_ffi_text, slug_ffi_text);
+typedef bool (*slug_ffi_set_resource_fn)(slug_ffi_call *, slug_ffi_text, void *);
+typedef bool (*slug_ffi_close_resource_fn)(slug_ffi_call *, size_t, slug_ffi_text);
 
 struct slug_ffi_host_api {
   uint32_t abi_major;
@@ -33,13 +36,17 @@ struct slug_ffi_host_api {
   uint32_t table_size;
   slug_ffi_argument_i64_fn argument_i64;
   slug_ffi_argument_f64_fn argument_f64;
+  slug_ffi_argument_resource_fn argument_resource;
   slug_ffi_set_i64_fn set_i64;
   slug_ffi_set_f64_fn set_f64;
   slug_ffi_set_error_fn set_error;
+  slug_ffi_set_resource_fn set_resource;
+  slug_ffi_close_resource_fn close_resource;
 };
 
 typedef int32_t (*slug_ffi_callback)(const slug_ffi_host_api *, slug_ffi_call *, void *);
 typedef void (*slug_ffi_module_destroy_fn)(void *);
+typedef void (*slug_ffi_resource_destroy_fn)(void *);
 
 typedef struct {
   uint32_t descriptor_size;
@@ -51,6 +58,12 @@ typedef struct {
 } slug_ffi_function_descriptor;
 
 typedef struct {
+  uint32_t descriptor_size;
+  slug_ffi_text name;
+  slug_ffi_resource_destroy_fn destroy_resource;
+} slug_ffi_resource_descriptor;
+
+typedef struct {
   uint32_t abi_major;
   uint32_t abi_minor;
   uint32_t descriptor_size;
@@ -58,6 +71,8 @@ typedef struct {
   slug_ffi_module_destroy_fn destroy_module;
   const slug_ffi_function_descriptor *functions;
   uint64_t function_count;
+  const slug_ffi_resource_descriptor *resources;
+  uint64_t resource_count;
 } slug_ffi_module_descriptor;
 
 typedef const slug_ffi_module_descriptor *(*slug_ffi_module_init_fn)(